refactor: move security contracts to output crate
